core: Lower all Revset related utilities to it
Move the pest grammar, the parser and all DSL related parts to the new crate.
Its needed so we can move the `WorkspaceName` and `WorkspaceNameBuf` newtypes for the `WorkingCopyStore`
trait. See the next patch.

Part of #6284
