+++ lib/core/Cargo.toml
@@ -21,10 +21,13 @@ etcetera = { workspace = true }
 futures = { workspace = true }
 itertools = { workspace = true }
 jj-core-proc-macros = { workspace = true }
+pest = { workspace = true }
+pest_derive = { workspace = true }
 pollster = { workspace = true }
 ref-cast = { workspace = true }
 serde = { workspace = true }
 smallvec = { workspace = true }
+strsim = { workspace = true }
 tempfile = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
